feat: add role-based dashboard and dummy data
